sgld, clase base y decaimiento ratio de aprendizaje
